[feature] add homepage view

+++ Models/TornUser.cs
@@ -2,7 +2,7 @@
 
 public sealed record TornUser(int Level, int Player_Id, string Name, TornUserState Status, TornUserActivity Last_Action, TornUserFaction Faction) { }
 
-public sealed record TornUserState(string Description, string State, int Until)
+public sealed record TornUserState(string Description, string State, ulong Until)
 {
     public static implicit operator UserState(TornUserState state) => Enum.Parse<UserState>(state.State);
 }
